stubbable view module.exports.stub. Useful for unit tests.

+++ index.js
@@ -1,4 +1,7 @@
 module.exports = function (pid) {
+  if (module.exports.stub !== module.exports) {
+      return module.exports.stub.apply(this, arguments);
+  }
   try {
     return process.kill(pid,0)
   }
@@ -6,3 +9,4 @@ module.exports = function (pid) {
     return e.code === 'EPERM'
   }
 }
+module.exports.stub = module.exports;
